+++ src/arr_o_gpu/module/method/function/slicing/slicing.rs
@@ -158,7 +158,7 @@ impl ArrOgpuModule {
         let arr = GpuArray {
             length: output_shape.iter().product::<u32>() as usize,
             module: Arc::new(self.clone()),
-            pointer: (output_allocate.1 as usize, output_allocate.2 as usize),
+            pointer: (output_allocate.1, output_allocate.2),
             stride: output_stride,
             shape: output_shape,
             space_type: output_allocate.0,
